Removed interface QaRecommendationService
