+++ webcurator-core/src/test/java/org/webcurator/core/rules/QaRecommendationServiceImplTest.java
@@ -5,15 +5,15 @@ import org.junit.Ignore;
 import org.junit.Test;
 import org.webcurator.test.BaseWCTTest;
 
-public class QaRecommendationServiceImplTest extends BaseWCTTest<QaRecommendationServiceImpl> {
+public class QaRecommendationServiceImplTest extends BaseWCTTest<QaRecommendationService> {
     public QaRecommendationServiceImplTest() {
-        super(QaRecommendationServiceImpl.class, "", false);
+        super(QaRecommendationService.class, "", false);
 
     }
 
     @Before
     public void init() {
-        testInstance = new QaRecommendationServiceImpl();
+        testInstance = new QaRecommendationService();
 
 //        File fileRules = new File("src/main/resources/rules.drl");
         testInstance.setRulesFileName("rules.drl");
